+++ Dockerfile
@@ -1,5 +1,5 @@
 #FROM driplineorg/dripline-python:v5.0.0
-FROM dripline-python:localCbi
+FROM dripline-python:test1
 
 COPY . /usr/local/src/dripline-python-plugin
 
@@ -11,5 +11,5 @@ RUN pip install .
 RUN apt-get update && apt-get install -y tini
 
 WORKDIR /
-#ENTRYPOINT ["/usr/local/src/dripline-python-plugin/sagebrush/entrypoint.sh"]
+ENTRYPOINT ["/usr/local/src/dripline-python-plugin/sagebrush/entrypoint.sh"]
 #CMD ["rsyslogd", "--no-daemon"]
