+++ dripline/extensions/__init__.py
@@ -5,4 +5,4 @@ from .plc import *
 from .admx_sensor_logger import *
 from .muxer_service import *
 from .multi_format import *
-
+from .modbus_service import *
