+++ dripline/extensions/muxer_service.py
@@ -4,6 +4,7 @@ A class to interface with the multiplexer aka muxer instrument
 
 from dripline.core import ThrowReply, Entity, calibrate
 from dripline.implementations import EthernetSCPIService
+from sagebrush.functions import piecewise_cal
 import math
 
 import logging
@@ -14,34 +15,7 @@ __all__.append('MuxerService')
 
 _all_calibrations = []
 
-#lin-log is log(x)
-#log-lin is log(y)
-def piecewise_cal(values_x, values_y, this_x, log_x=False, log_y=False):
-    if log_x:
-        logger.debug("doing log x cal")
-        values_x = [math.log(x) for x in values_x]
-        if this_x <= 0: 
-            logger.warning("invalid value for a log function")
-            return np.inf
-        this_x = math.log(this_x)
-    if log_y:
-        logger.debug("doing log y cal")
-        values_y = [math.log(y) for y in values_y]
-    try:
-        high_index = [i>this_x for i in values_x].index(True)
-    except ValueError:
-        high_index = -1
-        logger.warning("raw value is above the calibration range, extrapolating")
-        #raise dripline.core.DriplineValueError("raw value is likely above calibration range")
-    if high_index == 0:
-        high_index = 1
-        logger.warning("raw value is below the calibration range, extrapolating")
-        #raise dripline.core.DriplineValueError("raw value is below calibration range")
-    m = (values_y[high_index] - values_y[high_index - 1]) / (values_x[high_index] - values_x[high_index - 1])
-    to_return = values_y[high_index - 1] + m * (this_x - values_x[high_index - 1])
-    if log_y:
-        to_return = math.exp(to_return)
-    return to_return
+
 
 # PT 100
 def pt100_cal(resistance):
