--- sagebrush/temp.txt
@@ -1,7 +1,0 @@
-+----------------------------------+-------------------------+
-| channel_details.connection_name  |       queue.name        |
-+----------------------------------+-------------------------+
-| 10.0.1.4:36916 -> 10.0.1.83:5672 | sidecar-receiver-switch |
-| 10.0.1.4:39302 -> 10.0.1.83:5672 | my_store                |
-| 10.0.1.9:41812 -> 10.0.1.83:5672 | channel-1-switch        |
-+----------------------------------+-------------------------+
